CMake 3.1.0-rc1

+++ Tests/CTestTestMemcheck/testAddressSanitizer.cmake
@@ -1,15 +1,18 @@
-# this file simulates a program that has been built with thread sanitizer
+# this file simulates a program that has been built with address sanitizer
 # options
 
 message("ASAN_OPTIONS = [$ENV{ASAN_OPTIONS}]")
 string(REGEX REPLACE ".*log_path=\"([^\"]*)\".*" "\\1" LOG_FILE "$ENV{ASAN_OPTIONS}")
 message("LOG_FILE=[${LOG_FILE}]")
 
+# if we are not asked to simulate address sanitizer don't do it
+if(NOT "$ENV{ASAN_OPTIONS}]" MATCHES "simulate_sanitizer.1")
+  return()
+endif()
 # clear the log file
 file(REMOVE "${LOG_FILE}.2343")
 
-# create an error of each type of thread santizer
-# these names come from tsan_report.cc in llvm
+# create an example error from address santizer
 
 file(APPEND "${LOG_FILE}.2343"
 "=================================================================
